+++ src/teslamate_charge_importer/models.py
@@ -5,7 +5,6 @@ from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
     from datetime import datetime
-    from decimal import Decimal
 
 
 @dataclass(frozen=True)
@@ -17,9 +16,9 @@ class ChargingSession:
     duration_min: int | None
     start_battery_level: int | None
     end_battery_level: int | None
-    charge_energy_added_kwh: Decimal | None  # TO DO: convert all Decimal instances to float
-    charge_energy_used_kwh: Decimal | None
-    cost: Decimal | None
+    charge_energy_added_kwh: float | None
+    charge_energy_used_kwh: float | None
+    cost: float | None
     geofence_name: str | None = None
     short_address: str | None = None
 
@@ -29,8 +28,8 @@ class EnergyBucket:
     charging_process_id: int
     bucket_start: datetime
     bucket_end: datetime
-    kwh_added: Decimal
-    avg_kw: Decimal  # kWh in 5 minutes * 12
+    kwh_added: float
+    avg_kw: float  # kWh in 5 minutes * 12
 
 
 @dataclass(frozen=True)
